Fix unit tests, add null check for AppOfficialFQDN

+++ rest/EngineTests/UnitTest1.cs
@@ -60,7 +60,7 @@ namespace Tests
         return null;
       }
 
-      UInt32 CarrierInfo.GetCellID()
+      ulong CarrierInfo.GetCellID()
       {
         return 0;
       }
